refactor: Split nonagg.rs executor into focused modules (#2278)
Refactor the 874-line `nonagg.rs` file into a modular structure with
clear separation of concerns. This improves code organization and
maintainability while preserving all existing functionality.

## Module Structure

- `validation.rs` (133 lines): WHERE clause subquery validation
- `simd.rs` (66 lines): SIMD filtering optimization using Arrow
- `iterator.rs` (233 lines): Iterator-based execution for simple queries
- `materialized.rs` (514 lines): Materialized execution for complex queries
- `without_from.rs` (55 lines): SELECT without FROM clause execution
- `mod.rs` (34 lines): Module coordination and re-exports

## Changes

- Split monolithic 874-line file into 6 focused modules
- Each module handles a distinct execution strategy or concern
- Preserved all existing functionality and test coverage
- All executor tests pass without modification
- Clear documentation for each module's responsibility

## Benefits

- Better code organization and navigation
- Easier to understand execution strategies
- Simpler to maintain and extend individual strategies
- Reduced cognitive load when working with specific features

Closes #2278
